Added code and unit tests to only allow admins to access certain endpoints

+++ tvm/src/tvm/text_endpoints.py
@@ -1,4 +1,5 @@
 from main import app, get_db
+from authentication import *
 from fastapi import Depends, HTTPException, status, Response
 from sqlalchemy import Column, Integer, String, DateTime
 from sqlalchemy.ext.declarative import declarative_base
@@ -31,46 +32,55 @@ def read_category(category_id: int, db: Session = Depends(get_db)):
 
 # Create a new category
 @app.post("/categories/")
-def create_category(category_name: str, db: Session = Depends(get_db)):
+def create_category(category_name: str, db: Session = Depends(get_db), current_user: User = Depends(get_current_user)):
     """
     Create a new category of risk.
     """
-    category = db.get(Category, category_name)
-    if category:
-        raise HTTPException(status_code=400, detail="This category already exists.")
-    new_category = Category(text=category_name)
-    db.add(new_category)
-    db.commit()
-    db.refresh(new_category)
-    return Response(status_code=201, content=f"Category {new_category.text} successfully created.")
+    if current_user.role == "admin":
+        category = db.get(Category, category_name)
+        if category:
+            raise HTTPException(status_code=400, detail="This category already exists.")
+        new_category = Category(text=category_name)
+        db.add(new_category)
+        db.commit()
+        db.refresh(new_category)
+        return Response(status_code=201, content=f"Category {new_category.text} successfully created.")
+    else:
+        raise HTTPException(status_code=403, detail="You are not allowed to create a category.")
 
 # Update the category with the given ID
 @app.put("/categories/{category_id}")
-def update_category(category_id: int, category_name: str, db: Session = Depends(get_db)):
+def update_category(category_id: int, category_name: str, db: Session = Depends(get_db), current_user: User = Depends(get_current_user)):
     """
     Update a specific category of risk for the given ID.
     """
-    category = db.get(Category, category_id)
-    if not category:
-        return HTTPException(status_code=404, detail="Category not found.")
-    setattr(category, Category.text, category_name)
-    db.commit()
-    db.refresh(category)
-    return Response(status_code=200, content=f"Category {category_name} successfully updated.")
+    if current_user.role == "admin":
+        category = db.get(Category, category_id)
+        if not category:
+            return HTTPException(status_code=404, detail="Category not found.")
+        setattr(category, Category.text, category_name)
+        db.commit()
+        db.refresh(category)
+        return Response(status_code=200, content=f"Category {category_name} successfully updated.")
+    else:
+        raise HTTPException(status_code=403, detail="You are not allowed to update a category.")
 
 # Delete the category with the given ID
 @app.delete("/categories/{category_id}")
-def delete_category(category_id: int, db: Session = Depends(get_db)):
+def delete_category(category_id: int, db: Session = Depends(get_db), current_user: User = Depends(get_current_user)):
     """
     Delete the category of risk corresponding to the given ID.
     """
-    category = db.get(Category, category_id)
-    if not category:
-        return HTTPException(status_code=404, detail="Category not found.")
-    db.delete(category)
-    db.commit()
-    db.refresh(category)
-    return Response(status_code=204, content=f"Category {category} successfully deleted.")
+    if current_user.role == "admin":
+        category = db.get(Category, category_id)
+        if not category:
+            return HTTPException(status_code=404, detail="Category not found.")
+        db.delete(category)
+        db.commit()
+        db.refresh(category)
+        return Response(status_code=204, content=f"Category {category} successfully deleted.")
+    else:
+        raise HTTPException(status_code=403, detail="You are not allowed to delete a category.")
 
 # Get all subcategories
 @app.get("/subcategories/")
@@ -96,46 +106,55 @@ def read_subcategory(subcategory_id: int, db: Session = Depends(get_db)):
 
 # Create a new subcategory
 @app.post("/subcategories/")
-def create_subcategory(subcategory_name: str, db: Session = Depends(get_db)):
+def create_subcategory(subcategory_name: str, db: Session = Depends(get_db), current_user: User = Depends(get_current_user)):
     """
     Creates a new subcategory of risk willingness.
     """
-    subcategory = db.query(SubCategory).filter(SubCategory.name == subcategory_name).first()
-    if subcategory:
-        raise HTTPException(status_code=400, detail="This subcategory already exists.")
-    new_subcategory = SubCategory(text=subcategory_name)
-    db.add(new_subcategory)
-    db.commit()
-    db.refresh(new_subcategory)
-    return Response(status_code=201, content=f"Subcategory {subcategory_name} successfully created.")
+    if current_user.role == "admin":
+        subcategory = db.query(SubCategory).filter(SubCategory.name == subcategory_name).first()
+        if subcategory:
+            raise HTTPException(status_code=400, detail="This subcategory already exists.")
+        new_subcategory = SubCategory(text=subcategory_name)
+        db.add(new_subcategory)
+        db.commit()
+        db.refresh(new_subcategory)
+        return Response(status_code=201, content=f"Subcategory {subcategory_name} successfully created.")
+    else:
+        raise HTTPException(status_code=403, detail="You are not allowed to create a subcategory.")
 
 # Update the subcategory with the given ID
 @app.put("/subcategories/{subcategory_id}")
-def update_subcategory(subcategory_id: int, subcategory_name: str, db: Session = Depends(get_db)):
+def update_subcategory(subcategory_id: int, subcategory_name: str, db: Session = Depends(get_db), current_user: User = Depends(get_current_user)):
     """
     Update a specific subcategory of risk willingness for the given ID.
     """
-    subcategory = db.get(SubCategory, subcategory_id)
-    if not subcategory:
-        return HTTPException(status_code=404, detail="Subcategory not found.")
-    setattr(subcategory, SubCategory.text, subcategory_name)
-    db.commit()
-    db.refresh(subcategory)
-    return Response(status_code=200, content=f"Subcategory {subcategory_name} successfully updated.")
+    if current_user.role == "admin":
+        subcategory = db.get(SubCategory, subcategory_id)
+        if not subcategory:
+            return HTTPException(status_code=404, detail="Subcategory not found.")
+        setattr(subcategory, SubCategory.text, subcategory_name)
+        db.commit()
+        db.refresh(subcategory)
+        return Response(status_code=200, content=f"Subcategory {subcategory_name} successfully updated.")
+    else:
+        raise HTTPException(status_code=403, detail="You are not allowed to update a subcategory.")
 
 # Delete the subcategory with the given ID
 @app.delete("/subcategories/{subcategory_id}")
-def delete_category(subcategory_id: int, db: Session = Depends(get_db)):
+def delete_category(subcategory_id: int, db: Session = Depends(get_db), current_user: User = Depends(get_current_user)):
     """
     Delete the subcategory of risk willingness corresponding to the given ID.
     """
-    subcategory = db.get(SubCategory, subcategory_id)
-    if not subcategory:
-        return HTTPException(status_code=404, detail="Subcategory not found.")
-    db.delete(subcategory)
-    db.commit()
-    db.refresh(subcategory)
-    return Response(status_code=204, content=f"Subcategory {subcategory} successfully deleted.")
+    if current_user.role == "admin":
+        subcategory = db.get(SubCategory, subcategory_id)
+        if not subcategory:
+            return HTTPException(status_code=404, detail="Subcategory not found.")
+        db.delete(subcategory)
+        db.commit()
+        db.refresh(subcategory)
+        return Response(status_code=204, content=f"Subcategory {subcategory} successfully deleted.")
+    else:
+        raise HTTPException(status_code=403, detail="You are not allowed to delete a subcategory.")
 
 # Gets all advice texts
 @app.get("/advisorytexts/")
@@ -161,43 +180,52 @@ def read_text(catid: int, subid: int, db: Session = Depends(get_db)):
 
 # Create new advice text
 @app.post("/advisorytexts/")
-def create_text(advice_text: str, db: Session = Depends(get_db)):
+def create_text(advice_text: str, db: Session = Depends(get_db), current_user: User = Depends(get_current_user)):
     """
     Creates a new advisory text template.
     """
-    advice = db.query(AdvisoryText).filter(AdvisoryText.text == advice_text).first()
-    if advice:
-        raise HTTPException(status_code=400, detail="This text already exists.")
-    new_advice = AdvisoryText(text=advice_text)
-    db.add(new_advice)
-    db.commit()
-    db.refresh(new_advice)
-    return Response(status_code=201, content=f"AdvisoryText {new_advice.text} successfully created.")
+    if current_user.role == "admin":
+        advice = db.query(AdvisoryText).filter(AdvisoryText.text == advice_text).first()
+        if advice:
+            raise HTTPException(status_code=400, detail="This text already exists.")
+        new_advice = AdvisoryText(text=advice_text)
+        db.add(new_advice)
+        db.commit()
+        db.refresh(new_advice)
+        return Response(status_code=201, content=f"AdvisoryText {new_advice.text} successfully created.")
+    else:
+        raise HTTPException(status_code=403, detail="You are not allowed to create an advisory text.")
 
 # Edits the advice text with the given ID
 @app.put("/advisorytexts/catid={category_id}&subid={subcategory_id}")
-def update_text(catid: int, subid: int, advice_text: str, db: Session = Depends(get_db)):
+def update_text(catid: int, subid: int, advice_text: str, db: Session = Depends(get_db), current_user: User = Depends(get_current_user)):
     """
     Updates an existing advisory text template for a given category and subcategory.
     """
-    advisorytext = db.query(AdvisoryText).filter(AdvisoryText.category_id == catid && AdvisoryText.subcategory_id == subid).first()
-    if not advisorytext:
-        return HTTPException(status_code=404, detail="Advice text not found.")
-    setattr(advisorytext, AdvisoryText.text, advice_text)
-    db.commit()
-    db.refresh(advisorytext)
-    return Response(status_code=200, content=f"Advisory Text {advice_text} successfully updated.")
+    if current_user.role == "admin":
+        advisorytext = db.query(AdvisoryText).filter(AdvisoryText.category_id == catid && AdvisoryText.subcategory_id == subid).first()
+        if not advisorytext:
+            return HTTPException(status_code=404, detail="Advice text not found.")
+        setattr(advisorytext, AdvisoryText.text, advice_text)
+        db.commit()
+        db.refresh(advisorytext)
+        return Response(status_code=200, content=f"Advisory Text {advice_text} successfully updated.")
+    else:
+        raise HTTPException(status_code=403, detail="You are not allowed to update an advisory text.")
 
 # Deletes the advice text with the given ID
 @app.delete("/advisorytexts/catid={category_id}&subid={subcategory_id}")
-def delete_text(catid: int, subid: int, db: Session = Depends(get_db)):
+def delete_text(catid: int, subid: int, db: Session = Depends(get_db), current_user: User = Depends(get_current_user)):
     """
     Deletes a specific advisory text template for a given category and subcategory.
     """
-    advisorytext = db.query(AdvisoryText).filter(AdvisoryText.category_id == catid && AdvisoryText.subcategory_id == subid).first()
-    if not advisorytext:
-        return HTTPException(status_code=404, detail="Advice text not found.")
-    db.delete(advisorytext)
-    db.commit()
-    db.refresh(advisorytext)
-    return Response(status_code=204, content=f"Subcategory {advisorytext} successfully deleted.")
+    if current_user.role == "admin":
+        advisorytext = db.query(AdvisoryText).filter(AdvisoryText.category_id == catid && AdvisoryText.subcategory_id == subid).first()
+        if not advisorytext:
+            return HTTPException(status_code=404, detail="Advice text not found.")
+        db.delete(advisorytext)
+        db.commit()
+        db.refresh(advisorytext)
+        return Response(status_code=204, content=f"Subcategory {advisorytext} successfully deleted.")
+    else:
+        raise HTTPException(status_code=403, detail="You are not allowed to delete an advisory text.")
